refactor: consolidate duplicated types between pipeline-engine and agent-runtime (#282)
