+++ packages/pipeline-engine/src/executor/phase-executor.ts
@@ -6,48 +6,31 @@
  * these to concrete types without type incompatibilities.
  */
 
-import type { Logger, TokenRecord } from '../types.js';
-import type { CheckpointManager } from '../checkpoint/checkpoint.js';
-import type { IssueProgressWriter } from '../progress/progress.js';
-
 /** Cross-cutting services used by every phase. */
-export interface PhaseServices {
-  launcher: {
-    launch(invocation: { agent: string; contextPath: string; outputPath: string; timeout?: number }): Promise<{ success: boolean; exitCode: number | null; timedOut: boolean; duration: number; stdout: string; stderr: string; tokenUsage: unknown; outputPath: string; outputExists: boolean; error?: string }>;
-  };
-  retryExecutor: {
-    executeWithRetry<T>(fn: () => Promise<T>, maxRetries?: number): Promise<T>;
-  };
-  tokenTracker: {
-    record(agent: string, phase: number, tokens: number): void;
-    getTotal(): number;
-  };
-  contextBuilder: {
-    build(params: Record<string, unknown>): Promise<string>;
-  };
-  resultParser: {
-    parse(outputPath: string): Promise<unknown>;
-  };
-  logger: Logger;
-}
+export type PhaseServices = {
+  launcher: any;
+  retryExecutor: any;
+  tokenTracker: any;
+  contextBuilder: any;
+  resultParser: any;
+  logger: any;
+};
 
 /** I/O and persistence dependencies. */
-export interface PhaseIO {
+export type PhaseIO = {
   progressDir: string;
-  progressWriter: IssueProgressWriter;
-  checkpoint: CheckpointManager;
-  commitManager: {
-    commitPhase(message: string): Promise<string | null>;
-  };
-}
+  progressWriter: any;
+  checkpoint: any;
+  commitManager: any;
+};
 
 /** Callbacks injected by the orchestrator. */
-export interface PhaseCallbacks {
-  recordTokens: (agent: string, tokens: { input?: number; output?: number; total: number }) => void;
+export type PhaseCallbacks = {
+  recordTokens: (agent: string, tokens: any) => void;
   checkBudget: () => void;
   updateProgress: () => Promise<void>;
-  setPR?: (pr: { number: number; url: string }) => void;
-}
+  setPR?: (pr: any) => void;
+};
 
 /**
  * All dependencies and shared state needed by a phase during execution.
